% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primertools-package.R
\docType{package}
\name{primertools}
\alias{primertools-package}
\alias{primertools}
\title{primertools: Utilities for PRIMER-ready CSVs}
\description{
Helper functions for writing PRIMER import files and related utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anhsmith.github.io/primertools/}
  \item Report bugs at \url{https://github.com/anhsmith/primertools/issues}
}

}
\author{
\strong{Maintainer}: Adam Smith \email{anhsmith@gmail.com}

}
\keyword{internal}
