% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{read_permanova_components_of_variation}
\alias{read_permanova_components_of_variation}
\title{Read PRIMER PERMANOVA "Estimates of components of variation" from a text output file}
\usage{
read_permanova_components_of_variation(path, encoding = "UTF-8")
}
\arguments{
\item{path}{Path to a PRIMER PERMANOVA text output file.}

\item{encoding}{File encoding passed to \code{\link[=readLines]{readLines()}}. Default \code{"UTF-8"}.}
}
\value{
A data frame containing estimates of components of variation.
}
\description{
Parses the \strong{Estimates of components of variation} section from a PRIMER
PERMANOVA text output file into a data frame.
}
\examples{
\dontrun{
covs <- read_permanova_components_of_variation("PM_SiteCycle.txt")
covs
}
}
